etc: Update link for EPS history data
The EPS has a feature to download telemetry stored in its storage.
However, since the download format is variable-length and complex,
parsing it on the Yamcs side is challenging. Therefore, we use a
method where the data is first parsed with a Python script and then
forwarded to a dedicated port as telemetry. Additionally, the stored
telemetry includes historical timestamps (64-bit float), so a
configuration to associate it as a "Generation time" will also be
added. This Python script will be added to the repository later.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>